Move FTIs to contents folder
