+++ ftw/contacts/browser/contactfolder.py
@@ -1,4 +1,4 @@
-from ftw.contacts.contact import IContact
+from ftw.contacts.contents.contact import IContact
 from ftw.contacts.utils import ALPHABET
 from ftw.contacts.utils import AlphabeticLetters
 from ftw.contacts.utils import normalized_first_letter
