+++ ftw/contacts/profiles/default/types/ftw.contacts.Member.xml
@@ -12,7 +12,7 @@
     <property name="allowed_content_types" />
 
   <!-- schema interface -->
-  <property name="schema">ftw.contacts.member.IMember</property>
+  <property name="schema">ftw.contacts.contents.member.IMember</property>
 
   <property name="behaviors">
       <element value="plone.app.content.interfaces.INameFromTitle" />
@@ -20,7 +20,7 @@
   </property>
 
   <!-- class used for content items -->
-  <property name="klass">ftw.contacts.member.Member</property>
+  <property name="klass">ftw.contacts.contents.member.Member</property>
 
   <!-- add permission -->
   <property name="add_permission">ftw.contacts.AddMember</property>
