+++ ftw/contacts/tests/test_member_type.py
@@ -1,6 +1,6 @@
 from ftw.builder import Builder
 from ftw.builder import create
-from ftw.contacts.member import IMember
+from ftw.contacts.contents.member import IMember
 from ftw.contacts.testing import FTW_CONTACTS_FUNCTIONAL_TESTING
 from ftw.testbrowser import browsing
 from plone.api.content import delete
